Добавлено: валидация входных данных server > controller > categories > all

+++ server/controllers/categories/remove.js
@@ -11,14 +11,21 @@ const log = SimpleNodeLogger.createSimpleLogger(opts);
 
 
 module.exports.remove = async(req, res) => {
-    try {
-        let categories = await Categories.deleteOne({ _id: req.body.id })
-        log.info(`Успешное удаление категории`);
-        res.status(200).json({
-            message: 'delete-success'
-        })
-    } catch (e) {
-        log.warn(`Неудачная попытка удаления категории (ошибка обращения к базе)`);
-        res.status(500).json(e)
+    if (typeof(req.body.id) === 'string') {
+        try {
+            let categories = await Categories.deleteOne({ _id: req.body.id })
+            log.info(`Успешное удаление категории`);
+            res.status(200).json({
+                message: 'delete-success'
+            })
+        } catch (e) {
+            log.warn(`Неудачная попытка удаления категории (ошибка обращения к базе)`);
+            res.status(500).json(e)
+        }
+    } else {
+        log.warn(`Ошибка удаления категории. 
+                      Данные не прошли валидацию!`);
+        res.status(500).json('Error server. Data did not pass verification')
     }
+
 }
